Add files via upload
This is my first sales dashboard project, built in Excel 2010. It shows sales data of Indian customers and helps to understand sales trends and patterns in a simple way.

Because it was made in Excel 2010, the design is quite basic, but this project is very special to me. It marks the beginning of my journey in learning data visualization and analysis, and I enjoy creating dashboards that make information easy to understand.